Fix type for ID in Postgres

+++ bot/infrastructure/storage_postgres.py
@@ -74,7 +74,7 @@ class StoragePostgres(Storage):
                     CREATE TABLE IF NOT EXISTS users
                     (
                         id SERIAL PRIMARY KEY,
-                        telegram_id INTEGER NOT NULL UNIQUE,
+                        telegram_id BIGINT NOT NULL UNIQUE,
                         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                         state TEXT DEFAULT NULL,
                         order_json TEXT DEFAULT NULL
